service: Increase default inactivity timeout to 60s
instead of 5 seconds. 5s is very short amount of time when system is very
loaded and during startup of a desktop session. There is not harm in
waiting 60s anyway. Also, the default value should be longer than the
agent registeration timeout, which is currently 100ms but will be increased
to 20s in a following patch.