Re-add gclue_accuracy_level_get_string()
This function was being generated when we were using autotools because it
was part of our template. We unintentially dropped it when we ported to
meson's gnome.mkenums_simple() as that uses it's own template that doesn't
include this function.

This broke ABI in 2.5.0 and hence we now re-add it, implementing it
manually.

Fixes #90.