wifi: Tell wpa_supplicant to keep scanning
It seems by default, wpa_supplicant won't keep a list of BSS around but
only if it has done a scan recently. This means we currently end up with
very bad accuracy since we only send out the ESSID of the currently
connected WiFi (if any).

To fix this, we'll now tell wpa_supplicant to start a scan when the source is
started. We keep doing that continuously if accuracy is street level or higher,
but wait 5 minutues between scans if requested accuracy is lower than that.

Fixes #91.