service: Drop geocode-glib copy
The geocode-glib copy is out-of-date and updating it is not as trivial as
it used to be since geocode-location module is not so independent of the
rest of the code any more. Also geocode-glib doesn't see much changes
these days, especially to the geocode-location module so the effort to
port the code, isn't really worth it.

We already copied over the needed code in a previous commit and now we
ditch our geode-glib internal copy and dependency.

Fixes #88.