lib: Avoid linking to system-installed library
When building geoclue, in some cases the already installed library is
prefered for linking rather than the one from the build directory. This
causes problems when an old version is already installed and one is
adding new API.

This commit partially adresses this issue. A similar issue appears to
exist when building the documentation though.