locator: fix logic error in the destructor
The current code calls the destructor of the base class *first* and *then* does
some cleanup, which is a no-no. Luckily it doesn't crash but it creates a subtle
memory leak: the code tries to disconnect a signal from an object but since this
object is NULLed in the base class destructor, the disconnect will just print
a warning and leak the closure.