gclue-wifi: Implement a cache of MLS lookup results
This should avoid a significant amount of the HTTPS traffic to the MLS
website generated by immobile computers and laptops which can see the
same set of Wi-Fi networks all the time, modulo some noise.

Currently the cache can grow without bound. This could be remedied in
future, either by putting size limits on the cache and expiring by
least-recently used (LRU caching); and/or by dropping the cache when low
on memory (as per `GMemoryMonitor`).

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #114