gclue-wifi: Cancel existing scan timeout before scheduling another
Since the state of wpa_supplicant is out of our control, `ScanDone`
signals could be received at any time. Previously, we’d schedule another
scan timeout for each one, without cancelling any previously-scheduled
timeout. This could result in multiple timeouts in flight, but with the
`GClueWifi` code only holding an ID for the most recent one.

If the `GClueWifi` object was then destroyed, it was possible for the
timeout callbacks to still be invoked, with a dangling pointer to the
`GClueWifi`, resulting in a crash.

Fix that by cancelling any in-flight timeout before scheduling another.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #126