wifi,mozilla: Fix maximum SSID length handling
An SSID can be 32 bytes but the `MAX_SSID_LEN` macro is used to declare
the size of the buffers including the nul byte. This causes excessive
`variant_to_string: assertion 'len < max_len' failed` errors if you
happen to have a network with a full 32 byte SSID. This can often occur
for APs that advertise a hidden SSID by setting it to all zeros.

Conversely, the `BSSID_LEN` and `BSSID_STR_LEN` macros were sized to
include the nul byte. Rework the usages so that the length macros are
consistently in terms of string length and not buffer length.

Fixes: #135